Update rotkey lookup payload
