+++ queryregistry/identity/sessions/models.py
@@ -3,15 +3,35 @@
 from __future__ import annotations
 
 from collections.abc import Awaitable, Callable
-from typing import TypedDict
+from typing import Any, TypedDict
+
+from pydantic import BaseModel, ConfigDict, field_validator
 
 from queryregistry.models import DBResponse
 
 __all__ = [
+  "GuidParams",
   "SecuritySnapshotCallable",
   "SecuritySnapshotRecord",
   "SecuritySnapshotRequestPayload",
 ]
+
+
+def _normalize_uuid(value: Any) -> str:
+  return str(value)
+
+
+class GuidParams(BaseModel):
+  """Generic payload carrying a user guid."""
+
+  model_config = ConfigDict(extra="forbid")
+
+  guid: str
+
+  @field_validator("guid")
+  @classmethod
+  def _normalize_guid(cls, value: Any) -> str:
+    return _normalize_uuid(value)
 
 
 class SecuritySnapshotRequestPayload(TypedDict):
